move errorHandler to its own file
