+++ rest-api/recipe/controllers/recipe-controllers.js
@@ -1,4 +1,5 @@
 const dbRecipe = require('../models/recipe-models');
+const errorHandler = require("../middlewares/errorHandler");
 
 module.exports = {
   getRecipes,
@@ -45,17 +46,4 @@ async function addRecipe(req, res) {
       error: errorHandler(error)
     });
   }
-}
-
-function errorHandler(error) {
-  switch (error.code) {
-    case '23505':
-      const matchedString = error.detail.match(/\(([^()]+)\)/g);
-      const cleanString = matchedString[1].replace(/([()])/g, "'");
-      return `The value you entered for ${cleanString} already exists`;
-    case '23502':
-      return `No value was entered for '${error.column}'`
-    default:
-      return error;
-  } 
 }
